test and fix for CommonServiceLocator adapter WRT non-keyed request
The CommonServiceLocator specifies that calling its non-keyed
interface will call the keyed interface with a value of key == null.
It also states that when a key is passed as null, it needs to be
handled as returned the 'default' instance.  Because of this, we
need to handle the key == null case specifically.  Without this
change, the newly added test fails as it attempts to find a binding
with a name of "" (empty string).